+++ src/main/java/com/enesd/myshelfbackend/repository/elasticsearch/custom/IMediaContentDocumentRepositoryCustom.java
@@ -5,7 +5,7 @@ import com.enesd.myshelfbackend.model.documents.MediaContentDocument;
 import java.util.List;
 
 public interface IMediaContentDocumentRepositoryCustom {
-    List<Integer> findAllIds();
+    List<Long> findAllIds();
 
     List<MediaContentDocument> findWithSearchTerm(String searchTerm);
 }
